chg: rename lib/filepath to lib/fs
